Add files via upload
version 2 of the mod
this version does not start the httpd server automatically.
instead, user has to Double-click the [Power] button on the camera. Then user should see a top light cycle green-orange-green to indicate it's working now.
this is for users who still want to use the BT mode when required. Because if httpd server is running , BT mode does not work.